+++ lib/zip_name_locate.c
@@ -49,6 +49,7 @@ zip_name_locate(zip_t *za, const char *fname, zip_flags_t flags) {
 zip_int64_t
 _zip_name_locate(zip_t *za, const char *fname, zip_flags_t flags, zip_error_t *error) {
     int (*cmp)(const char *, const char *);
+    zip_string_t *str = NULL;
     const char *fn, *p;
     zip_uint64_t i;
 
@@ -60,7 +61,17 @@ _zip_name_locate(zip_t *za, const char *fname, zip_flags_t flags, zip_error_t *e
         return -1;
     }
 
-    if (flags & (ZIP_FL_NOCASE | ZIP_FL_NODIR | ZIP_FL_ENC_CP437)) {
+    if ((flags & (ZIP_FL_ENC_UTF_8 | ZIP_FL_ENC_RAW)) == 0 && fname[0] != '\0') {
+        if ((str = _zip_string_new((const zip_uint8_t *)fname, strlen(fname), flags, error)) == NULL) {
+            return -1;
+        }
+        if ((fname = (const char *)_zip_string_get(str, NULL, 0, error)) == NULL) {
+            _zip_string_free(str);
+            return -1;
+        }
+    }
+
+    if (flags & (ZIP_FL_NOCASE | ZIP_FL_NODIR | ZIP_FL_ENC_RAW | ZIP_FL_ENC_STRICT)) {
         /* can't use hash table */
         cmp = (flags & ZIP_FL_NOCASE) ? strcasecmp : strcmp;
 
@@ -79,14 +90,18 @@ _zip_name_locate(zip_t *za, const char *fname, zip_flags_t flags, zip_error_t *e
 
             if (cmp(fname, fn) == 0) {
                 _zip_error_clear(error);
+                _zip_string_free(str);
                 return (zip_int64_t)i;
             }
         }
 
         zip_error_set(error, ZIP_ER_NOENT, 0);
+        _zip_string_free(str);
         return -1;
     }
     else {
-        return _zip_hash_lookup(za->names, (const zip_uint8_t *)fname, flags, error);
+        zip_int64_t ret = _zip_hash_lookup(za->names, (const zip_uint8_t *)fname, flags, error);
+        _zip_string_free(str);
+        return ret;
     }
 }