+++ NEWS.md
@@ -3,6 +3,7 @@
 
 * Add `zip_source_is_seekable()`.
 * Improve compatibility with WinAES.
+* Fix encoding handling in `zip_name_locate()`.
 
 1.8.0 [2021-06-18]
 ==================